Add ReadGenericWithUnstructured
And change ApplyDirectly + StaticResourceController to use it.

This allows StaticResourceController.RelatedObjects() to process unknown
types as Unstructured.
