+++ pkg/operator/resource/resourceapply/generic.go
@@ -10,40 +10,19 @@ import (
 	rbacv1 "k8s.io/api/rbac/v1"
 	storagev1 "k8s.io/api/storage/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apiextensionsv1beta1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1"
 	apiextensionsclient "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/kubernetes/scheme"
 	corev1client "k8s.io/client-go/kubernetes/typed/core/v1"
 	migrationv1alpha1 "sigs.k8s.io/kube-storage-version-migrator/pkg/apis/migration/v1alpha1"
 	migrationclient "sigs.k8s.io/kube-storage-version-migrator/pkg/clients/clientset"
 
-	"github.com/openshift/api"
-
 	"github.com/openshift/library-go/pkg/operator/events"
+	"github.com/openshift/library-go/pkg/operator/resource/resourceread"
 	"github.com/openshift/library-go/pkg/operator/v1helpers"
 )
-
-var (
-	genericScheme = runtime.NewScheme()
-	genericCodecs = serializer.NewCodecFactory(genericScheme)
-	genericCodec  = genericCodecs.UniversalDeserializer()
-)
-
-func init() {
-	utilruntime.Must(api.InstallKube(genericScheme))
-	utilruntime.Must(apiextensionsv1beta1.AddToScheme(genericScheme))
-	utilruntime.Must(apiextensionsv1.AddToScheme(genericScheme))
-	utilruntime.Must(migrationv1alpha1.AddToScheme(genericScheme))
-	utilruntime.Must(admissionregistrationv1.AddToScheme(genericScheme))
-	// TODO: remove once openshift/api/pull/929 is merged
-	utilruntime.Must(policyv1.AddToScheme(genericScheme))
-}
 
 type AssetFunc func(name string) ([]byte, error)
 
@@ -112,7 +91,7 @@ func ApplyDirectly(ctx context.Context, clients *ClientHolder, recorder events.R
 			ret = append(ret, result)
 			continue
 		}
-		requiredObj, err := decode(objBytes)
+		requiredObj, err := resourceread.ReadGenericWithUnstructured(objBytes)
 		if err != nil {
 			result.Error = fmt.Errorf("cannot decode %q: %v", file, err)
 			ret = append(ret, result)
@@ -254,7 +233,7 @@ func DeleteAll(ctx context.Context, clients *ClientHolder, recorder events.Recor
 			ret = append(ret, result)
 			continue
 		}
-		requiredObj, err := decode(objBytes)
+		requiredObj, err := resourceread.ReadGenericWithUnstructured(objBytes)
 		if err != nil {
 			result.Error = fmt.Errorf("cannot decode %q: %v", file, err)
 			ret = append(ret, result)
@@ -389,20 +368,4 @@ func (c *ClientHolder) secretsGetter() corev1client.SecretsGetter {
 		return c.kubeClient.CoreV1()
 	}
 	return v1helpers.CachedSecretGetter(c.kubeClient.CoreV1(), c.kubeInformers)
-}
-
-func decode(objBytes []byte) (runtime.Object, error) {
-	// Try to get a typed object first
-	typedObj, _, decodeErr := genericCodec.Decode(objBytes, nil, nil)
-	if decodeErr == nil {
-		return typedObj, nil
-	}
-
-	// Try unstructured, hoping to recover from "no kind XXX is registered for version YYY"
-	unstructuredObj, _, err := scheme.Codecs.UniversalDecoder().Decode(objBytes, nil, &unstructured.Unstructured{})
-	if err != nil {
-		// Return the original error
-		return nil, decodeErr
-	}
-	return unstructuredObj, nil
 }
