+++ pkg/operator/resource/resourceapply/generic_test.go
@@ -4,27 +4,10 @@ import (
 	"context"
 	"testing"
 
-	"github.com/davecgh/go-spew/spew"
-
 	"k8s.io/client-go/kubernetes/fake"
 
 	"github.com/openshift/library-go/pkg/operator/events"
 )
-
-func TestApplyDirectly(t *testing.T) {
-	requiredObj, gvk, err := genericCodec.Decode([]byte(`apiVersion: v1
-kind: Namespace
-metadata:
-  name: openshift-apiserver
-  labels:
-    openshift.io/run-level: "1"
-`), nil, nil)
-	t.Log(spew.Sdump(requiredObj))
-	t.Log(spew.Sdump(gvk))
-	if err != nil {
-		t.Fatal(err)
-	}
-}
 
 func TestApplyDirectlyUnhandledType(t *testing.T) {
 	fakeClient := fake.NewSimpleClientset()
